// client/src/types/types.ts
export interface GalleryTypes {
  id: number;
  title: string;
  img: string;
}
export interface DestinationTypes {
  id: number;
  destinationName: string;
  destinationImg: string;
}

export interface BlogTypes {
  id: string;
  imageUrl: string;
  title: string;
  text: string;
  author: string;
  category: string;
  createdAt: Date;
}